Rename parameter `async` in usages of clingo.Control.solve into `async_`
Python 3.7 introduced `async` as a new keyword. That broke the Python API of
Clingo since clingo.Control.solve used to have a parameter which was also
called `async`. That problem was resolved with the release of Clingo 5.4.0
which renamed the parameter `async` into `async_`. This commit renames the
parameter `async` into `async_` throughout all usages of clingo.Control.solve
in asprin.

This change should (hopefully) make asprin compatible with Clingo 5.4.0 and
Python 3.7.